[Silver III] Title: 주유소, Time: 20 ms, Memory: 2412 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -14,7 +14,7 @@ int main()
 	for (int i = 0; i < n - 1; i++) {
 		cin >> a;
 		Min = min(Min, a); //지금까지의 가장 싼 가격
-		sum += Min * len[i];
+		sum += (long long)Min * len[i];
 	}
 	cin >> a;
 	cout << sum;
